# perf: use filter instead of take in Patches::filter