--- a/Cargo.toml
+++ b/Cargo.toml
@@ -228,3 +228,7 @@ debug = true
 [profile.benchtest]
 inherits = "bench"
 debug-assertions = true
+
+[profile.samply]
+inherits = "release"
+debug = true