--- a/bench-vortex/Cargo.toml
+++ b/bench-vortex/Cargo.toml
@@ -61,6 +61,7 @@ tokio = { workspace = true, features = ["full"] }
 uuid = { workspace = true, features = ["v4"] }
 vortex = { workspace = true, features = ["object_store", "parquet"] }
 vortex-datafusion = { workspace = true }
+vortex-mask = { workspace = true }
 xshell = { workspace = true }
 
 [dev-dependencies]
@@ -98,5 +99,10 @@ harness = false
 
 [[bench]]
 name = "clickbench"
+test = false
+harness = false
+
+[[bench]]
+name = "sel_vec"
 test = false
 harness = false