@a10y 2025-01-27
Performance difference if we don't do the stride-by-4 overlap check:

<img width="657" alt="image" src="https://github.com/user-attachments/assets/01a4c305-487d-4763-b67b-48bd06a3e4e1" />

I'm not fully sure why selecitivity=0.1 is so much better without striding