@a10y 2025-01-28
<img width="569" alt="image" src="https://github.com/user-attachments/assets/1b27fb28-8cac-43f0-ab7e-c9be8ae4f09f" />

<img width="574" alt="image" src="https://github.com/user-attachments/assets/8fec3efa-720d-48d8-b2fa-1bcf0f9e18bd" />


Seems like we get a nice little boost for Q7/11 in Clickbench